tests green

+++ W3ChampionsStatisticService/Ladder/PlayOverviewHandler.cs
@@ -38,7 +38,7 @@ namespace W3ChampionsStatisticService.Ladder
             var winnerPlayerIds = players.Select(w => PlayerId.Create(w.battleTag)).ToList();
 
             var winnerIdCombined =
-                $"{string.Join("_", winnerPlayerIds.OrderBy(t => t.BattleTag).Select(t => $"{t.BattleTag}@{nextEvent.match.gateway}"))}_{nextEvent.match.gameMode}";
+                $"{string.Join("_", winnerPlayerIds.OrderBy(t => t.BattleTag).Select(t => $"{t.BattleTag}@{(int)nextEvent.match.gateway}"))}_{nextEvent.match.gameMode}";
 
             var winner = await _playerRepository.LoadOverview(winnerIdCombined)
                          ?? PlayerOverview.Create(
