+++ W3ChampionsStatisticService/Ladder/PlayerOverview.cs
@@ -12,7 +12,7 @@ namespace W3ChampionsStatisticService.Ladder
         {
             return new PlayerOverview
             {
-                Id = $"{string.Join("_", playerIds.OrderBy(t => t.BattleTag).Select(t => $"{t.BattleTag}@{gateWay}"))}_{gameMode}",
+                Id = $"{string.Join("_", playerIds.OrderBy(t => t.BattleTag).Select(t => $"{t.BattleTag}@{(int)gateWay}"))}_{gameMode}",
                 PlayerIds = playerIds,
                 GateWay = gateWay,
                 GameMode = gameMode
